requirements.txt
anthropic>=0.7.0
python-dotenv>=1.0.0
twilio>=8.0.0
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0
xlrd>=2.0.0
sqlalchemy>=2.0.0
duckdb>=0.9.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
scikit-learn>=1.3.0
scipy>=1.10.0
pyarrow>=12.0.0
fastparquet>=0.8.0
tabulate>=0.9.0
jinja2>=3.1.0
pyyaml>=6.0.0
jsonschema>=4.17.0
dateparser>=1.1.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.21.0 